+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
-All notable changes to the "luau-library-class-generator" extension will be documented in this file.
+All notable changes to the "luau-class-generator" extension will be documented in this file.
+
+## [v1.0.5]
+- Patched a bug where the license name configuration was read as undefined.
+- Added a command: `LCG: New File + Class`
+- Added an `@lib <libraryName>` parameter to class method comment details.
 
 ## [v1.0.1]
 - Added Method Generation Command
@@ -11,5 +16,4 @@ All notable changes to the "luau-library-class-generator" extension will be docu
 - Official Full Release
 
 ## [v0.0.1]
-
 - Initial release